feat: add json and logging feature. Return diagnostics instead of outputting.
